Update measure/factors error messages

+++ value/measures/views.py
@@ -54,15 +54,17 @@ def add(request):
 def edit(request, measure_id):
     measure = get_object_or_404(Measure, pk=measure_id)
     MeasureValueFormSet = inlineformset_factory(Measure, MeasureValue, fields=('description', 'order', 'color'), extra=0)
+    can_edit = (not measure.deliverables.exists() and not measure.meetings.exists())
     if request.method == 'POST':
         form = ChangeMeasureForm(request.POST, instance=measure)
         formset = MeasureValueFormSet(request.POST, instance=measure)
         if form.is_valid() and formset.is_valid():
             form.save()
-            formset.save()
-            for value in measure.measurevalue_set.all():
-                if not value.description.strip():
-                    value.delete()
+            if can_edit:
+                formset.save()
+                for value in measure.measurevalue_set.all():
+                    if not value.description.strip():
+                        value.delete()
             messages.success(request, u'The measure {0} was changed successfully.'.format(measure.name))
             return redirect(reverse('measures:index'))
         else:
@@ -70,22 +72,18 @@ def edit(request, measure_id):
     else:
         form = ChangeMeasureForm(instance=measure)
         formset = MeasureValueFormSet(instance=measure)
-    return render(request, 'measures/edit.html', { 'form' : form, 'formset' : formset })
+    return render(request, 'measures/edit.html', { 'form': form, 'formset': formset, 'can_edit': can_edit })
 
 @login_required
 @user_passes_test(lambda user: user.is_superuser)
 def delete(request, measure_id):
     measure = get_object_or_404(Measure, pk=measure_id)
-    measure_has_relations = measure.deliverables.exists()
-    if measure_has_relations:
-        associated_deliverables = measure.deliverables.values_list('name', flat=True)
-        str_associated_deliverables = ', '.join(associated_deliverables)
-        messages.warning(request, u'The measure {0} is associated with deliverables. Deleting this measure will lead to data loss. If you really want to delete this measure from the system, please delete the associated deliverables first: {1}.'.format(measure.name, str_associated_deliverables))
-    if request.method == 'POST' and not measure_has_relations:
+    can_delete = (not measure.deliverables.exists() and not measure.meetings.exists())
+    if request.method == 'POST' and can_delete:
         measure.delete()
         messages.success(request, u'The measure {0} was deleted successfully.'.format(measure.name))
         return redirect(reverse('measures:index'))
-    return render(request, 'measures/delete.html', { 'measure' : measure, 'can_delete': measure_has_relations })
+    return render(request, 'measures/delete.html', { 'measure': measure, 'can_delete': can_delete })
 
 @login_required
 @user_passes_test(lambda user: user.is_superuser)
